+++ src/tt_user.go
@@ -782,7 +782,7 @@ func xmit_object_report(i int, first_time bool) {
 
 		dedupe_remember(pp, save_tt_config_p.obj_xmit_chan)
 
-		tq_append(C.int(save_tt_config_p.obj_xmit_chan), TQ_PRIO_1_LO, pp)
+		tq_append(save_tt_config_p.obj_xmit_chan, TQ_PRIO_1_LO, pp)
 	} else {
 		ax25_delete(pp)
 	}
